+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr2 (development version)
 
+* `req_retry()` gives a clearer error if `after` returns the wrong type of 
+  value (#385).
+
 # httr2 1.0.0
 
 ## Function lifecycle